Move vpc commands under "exo networking vpc"
