+++ cmd/networking/networking.go
@@ -1,0 +1,19 @@
+// AI-generated by claude-opus-5 - not reviewed yet
+package networking
+
+import (
+	exocmd "github.com/exoscale/cli/cmd"
+	"github.com/spf13/cobra"
+)
+
+// NetworkingCmd is the root command for networking services.
+var NetworkingCmd = &cobra.Command{
+	Use:        "networking",
+	Short:      "Networking services management",
+	Aliases:    []string{"net"},
+	SuggestFor: []string{"network", "vpc"},
+}
+
+func init() {
+	exocmd.RootCmd.AddCommand(NetworkingCmd)
+}
