+++ cmd/networking/vpc/vpc_route_delete.go
