+++ cmd/networking/vpc/vpc_update_test.go
